+++ src/terraform_manager/entities/terraform.py
@@ -1,5 +1,6 @@
 from typing import Optional, List
 
+from terraform_manager.entities.workspace import Workspace
 from terraform_manager.terraform.locking import lock_or_unlock_workspaces
 from terraform_manager.terraform.versions import check_versions, patch_versions, \
     write_version_summary, group_by_version
@@ -15,21 +16,28 @@ class Terraform:
         *,
         workspace_names: Optional[List[str]] = None,
         blacklist: bool = False,
+        no_tls: bool = False,
         write_output: bool = False
     ):
         self.terraform_domain = terraform_domain
         self.organization = organization
         self.workspace_names = workspace_names
         self.blacklist = blacklist
+        self.no_tls = no_tls
         self.write_output = write_output
+        self._workspace_cache: Optional[List[Workspace]] = None
 
-        self.workspaces = fetch_all(
-            terraform_domain,
-            organization,
-            workspace_names=workspace_names,
-            blacklist=blacklist,
-            write_error_messages=write_output
-        )
+    @property
+    def workspaces(self) -> List[Workspace]:
+        if self._workspace_cache is None:
+            self._workspace_cache = fetch_all(
+                self.terraform_domain,
+                self.organization,
+                workspace_names=self.workspace_names,
+                blacklist=self.blacklist,
+                write_error_messages=self.write_output
+            )
+        return self._workspace_cache
 
     def lock_workspaces(self) -> bool:
         """
@@ -43,6 +51,7 @@ class Terraform:
             self.organization,
             self.workspaces,
             set_lock=True,
+            no_tls=self.no_tls,
             write_output=self.write_output
         )
 
@@ -58,6 +67,7 @@ class Terraform:
             self.organization,
             self.workspaces,
             set_lock=False,
+            no_tls=self.no_tls,
             write_output=self.write_output
         )
 
@@ -86,6 +96,7 @@ class Terraform:
             self.organization,
             self.workspaces,
             new_version=new_version,
+            no_tls=self.no_tls,
             write_output=self.write_output
         )
 
@@ -116,18 +127,20 @@ class Terraform:
             self.organization,
             self.workspaces,
             new_working_directory=new_working_directory,
+            no_tls=self.no_tls,
             write_output=self.write_output
         )
 
     def __repr__(self):
         return (
             "Terraform(domain={}, organization={}, workspaces=List[{}], blacklist={}, "
-            "write_output={})"
+            "no_tls={}, write_output={})"
         ).format(
             self.terraform_domain,
             self.organization,
             len(self.workspaces),
             self.blacklist,
+            self.no_tls,
             self.write_output
         )